src/app/use-cases/unread-notification.ts
import { Injectable } from "@nestjs/common";
import { NotificationsRepository } from "../repositories/notifications-repository";
import { NotificationNotFound } from "./errors/notification-not-fonund";
import { NotificationNotRead } from "./errors/notification-not-read";

interface UnreadNotificationRequest {
  notificationId: string;
}

type UnreadNotificationResponse = void;

@Injectable()
export class UnreadNotification {
  constructor(private notificationRepository: NotificationsRepository) {}

  async execute(
    request: UnreadNotificationRequest,
  ): Promise<UnreadNotificationResponse> {
    const { notificationId } = request;

    const notification = await this.notificationRepository.findById(
      notificationId,
    );

    if (!notification) {
      throw new NotificationNotFound();
    }

    const notificationNotRead = !notification.isRead();

    if (notificationNotRead) {
      throw new NotificationNotRead();
    }

    notification.unread();

    await this.notificationRepository.save(notification);
  }
}
